Bump flycheck dependency
We are using `flycheck-relevant-error-other-file-p`, which is in the
development version of Flycheck.  This is causing issue for people using
MELPA stable.

The fix is to depend on the unreleased 32 version, which will fail for
people using a mix of MELPA stable and unstable, but at least the error
will be less obscure than a void function.

Close #7.